[#1540] Many Changes and fixes
Shop Id on Categories is hard coded with the id 2, In prestashop, the id
of the root category can be something else than 2, it's just the default
category. For that reason, I also imported the category id linked with
the shop id. So we can set the right root category.

Made product variant export not delayed. Force export synchronuously,
it's not a huge deal to have it this way. It reduce the amount of error
caused by trying to export the same thing multiple times.

Use the extra price on value attributes instead of the weird new "price
impact field". Why duplicate what already exists...?